pyproject.toml
[tool.poetry]
name = "inference-service"
version = "0.1.0"
description = ""
authors = ["Arun Hegde <ashegde13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
loguru = "^0.7.2"
pydantic-settings = "^2.4.0"
scikit-learn = "^1.5.1"
wemake-python-styleguide = "^0.19.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
